fix warnings

+++ src/database/database_tools/implementation/photos_analyzer.cpp
@@ -112,7 +112,7 @@ void PhotosAnalyzerImpl::addPhotos(const std::vector<Photo::Id>& ids)
         {
             int progress = 0;
             loadTask->getProgressBar()->setMinimum(0);
-            loadTask->getProgressBar()->setMaximum(count);
+            loadTask->getProgressBar()->setMaximum(static_cast<int>(count));
 
             slice(ids.begin(), ids.end(), 200, [this, loadTask, &progress](auto first, auto last)
             {
@@ -120,7 +120,7 @@ void PhotosAnalyzerImpl::addPhotos(const std::vector<Photo::Id>& ids)
                     evaluate<std::vector<Photo::Data>(Database::IBackend &)>(m_database, [first, last](Database::IBackend& backend)
                 {
                     std::vector<Photo::Data> photos;
-                    photos.reserve(last - first);
+                    photos.reserve(static_cast<unsigned>(last - first));
 
                     std::transform(first, last, std::back_inserter(photos), [&backend](const Photo::Id& id) mutable
                     {
